requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
argparse==1.2.1
enum==0.4.4
grizzled-python==1.0.1
itsdangerous==0.24
lxml==3.3.5
prettytable==0.7.2
sae-python-dev==1.3.2
sqlcmd==0.7.1
wsgiref==0.1.2
